% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loop.est.obj.R
\name{confint.loopEst}
\alias{confint.loopEst}
\title{Confidence interval for LOOP estimate}
\usage{
\method{confint}{loopEst}(x, level = 0.95, alpha = 1 - level, ...)
}
\arguments{
\item{x}{LOOP estimate}
}
\description{
Confidence interval for LOOP estimate
}
