% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HTE_helper_functions.R
\name{getITE}
\alias{getITE}
\title{Extract unbiased individual treatment effect estimates from a LOOP object}
\usage{
getITE(obj)
}
\arguments{
\item{obj}{LOOP estimate, output from the `loop()` function}
}
\description{
Extract unbiased individual treatment effect estimates from a LOOP object
}
