% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loop_mean.R
\name{loop_mean}
\alias{loop_mean}
\title{Leave-One-Out Mean Imputation of Potential Outcomes}
\usage{
loop_mean(Y, Tr, Z, ...)
}
\arguments{
\item{Y}{A vector of observed outcomes.}

\item{Tr}{The treatment assignment vector.}

\item{Z}{A matrix of pre-treatment covariates. Ignored.}

\item{...}{ignored.}
}
\description{
This function imputes subject i's potential outcomes using the means of
other subjects' observed outcomes.
The resulting ATE estimate is equal to the Neyman difference-in-means
estimate; hence, this function is intended for illustrative purposes.
}
