% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p_loomi.R
\name{p_loomi}
\alias{p_loomi}
\title{Leave-One-Out Mean Imputation of Potential Outcomes in Paired Experiments}
\usage{
p_loomi(ordered, assigned, n_assigned, loo = TRUE, weighted_imp = FALSE)
}
\arguments{
\item{ordered}{A matrix of pair experimental data that his been processed by the \code{pair} function, with the treatment pair first.}

\item{assigned}{A matrix of pair experimental data that his been processed by the \code{pair} function.}

\item{n_assigned}{A matrix of pair experimental data cluster sizes that has been processed by the \code{pair} function.}

\item{loo}{If set to \code{TRUE}, leave-one-pair-out imputation should be used.}

\item{weighted_imp}{If set to \code{TRUE}, cluster sizes will be used as weights in imputation models.}
}
\description{
This function is used to impute potential outcomes for the \code{p_loop} function. Pairs are treated
as the unit. This function used with \code{p_loop} implements the LOO-MI estimator when \code{loo} is set to \code{TRUE} (the default).
If \code{loo} is set to \code{FALSE} and  \code{weighted_imp} is set to \code{TRUE}, (or all of the cluster sizes are the same),
then the point estimator returned by \code{p_loop} is equivalent to the difference in means estimator.
}
