% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p_loop.R
\name{p_loop}
\alias{p_loop}
\title{P-LOOP Estimator}
\usage{
p_loop(
  Y,
  Tr,
  Z = NULL,
  P,
  n = NULL,
  pred = p_rf_interp,
  returnFitInfo = FALSE,
  ...
)
}
\arguments{
\item{Y}{A vector of mean experimental outcomes for each unit.}

\item{Tr}{The treatment assignment vector.}

\item{Z}{A matrix or vector of covariates.}

\item{P}{A vector encoding the pair assignments. Observations with the same value of \code{P} will be assumed to be in the same pair.}

\item{n}{A vector encoding the cluster sizes if there are pairs of clusters rather than individuals.}

\item{pred}{The prediction algorithm used to impute potential outcomes. By default, this is \code{p_rf_interp}, which uses random forests and interpolates between methods \code{p_rf_v12} and \code{p_rf_po} that treat the pairs as units or treats the individuals as units when making predictions. Another option is \code{p_ols_interp}, which uses linear regression. User written imputation methods may be used as well.}

\item{returnFitInfo}{If set to \code{TRUE}, returns detailed information about the prediction algorithm fit.}

\item{...}{Arguments to be passed to the imputation method.}
}
\description{
Design-based estimator of the average treatment effect in pair randomized and paired cluster-randomized experiments. 
Given a set of experimental data (including observed outcome, treatment assignment, pair assignments, and covariates), uses a method \code{pred} to impute the potential outcomes for each observation, which are then used to estimate the average treatment effect.
}
