% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p_loop_var.R
\name{p_loop_var}
\alias{p_loop_var}
\title{Variance Estimate for the P-LOOP Estimator}
\usage{
p_loop_var(assigned, v1, v2, n_assigned)
}
\arguments{
\item{assigned}{A matrix of pair experimental data that has been processed by the \code{pair} function}

\item{v1}{A vector of the predicted treatment minus control differences for the pairs where the first unit is assigned to treatment.}

\item{v2}{A vector of the predicted treatment minus control differences for the pairs where the second unit is assigned to treatment.}

\item{n_assigned}{A matrix of cluster sizes that has been processed by the \code{pair} function
#' @export}
}
\description{
This function implements the variance estimate for the P-LOOP estimator and is used within \code{p_loop}.
}
