% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p_ranger_po.R
\name{p_ranger_po_OneRF}
\alias{p_ranger_po_OneRF}
\title{Ranger-Based Combined Random Forest Imputation of Potential Outcomes in Paired Experiments (Individuals as Units)}
\usage{
p_ranger_po_OneRF(
  ordered,
  assigned,
  n_assigned,
  weighted_imp = FALSE,
  ntree = 500 * exp(1),
  nodesize = 5,
  ...
)
}
\arguments{
\item{ordered}{A matrix of pair experimental data that his been processed by the \code{pair} function, with the treatment pair first.}

\item{assigned}{A matrix of pair experimental data that his been processed by the \code{pair} function.}

\item{n_assigned}{A matrix of pair experimental data cluster sizes that has been processed by the \code{pair} function.}

\item{weighted_imp}{If set to \code{TRUE}, cluster sizes will be used as weights in imputation models.}

\item{ntree}{Integer specifying the number of trees to grow. Default is `500*exp(1)`}

\item{nodesize}{Integer specifying the minimum node size. Default is 5.}
}
\description{
This function is used to impute potential outcomes for the \code{p_loop} function.
It functions similarly to p_rf_po(), but instead of fitting separate random forests
leaving out each pair with the `randomForest` package, it fits one random forest,
including the treatment indicator as a predictor and using the `ranger` package,
but ensures that the out-of-box predictions use trees that excluded both members
of each pair.
}
