% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p_rf_interp.R
\name{p_rf_interp}
\alias{p_rf_interp}
\title{Random Forest Imputation of Potential Outcomes in Paired Experiments (Interpolated)}
\usage{
p_rf_interp(ordered, assigned, n_assigned, reparm = TRUE, weighted_imp = F)
}
\arguments{
\item{ordered}{A matrix of pair experimental data that his been processed by the \code{pair} function, with the treatment pair first.}

\item{assigned}{A matrix of pair experimental data that his been processed by the \code{pair} function.}

\item{n_assigned}{A matrix of pair experimental data cluster sizes that has been processed by the \code{pair} function.}

\item{reparm}{If set to \code{TRUE}, covariates will be parameterized as the pairwise differences and means for each covariate.}

\item{weighted_imp}{If set to \code{TRUE}, cluster sizes will be used as weights in imputation models.}
}
\description{
This function is used to impute potential outcomes for the \code{p_loop} function. 
Leaves out each pair and imputes its potential outcomes using random forests
on the remaining pairs. This function interpolates between the \code{p_rf_v12} and \code{p_rf_po} methods.
}
