% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p_rf_po.R
\name{p_rf_po}
\alias{p_rf_po}
\title{Random Forest Imputation of Potential Outcomes in Paired Experiments}
\usage{
p_rf_po(ordered, assigned, n_assigned)
}
\arguments{
\item{ordered}{A matrix of pair experimental data that his been processed by the \code{pair} function, with the treatment pair first.}

\item{assigned}{A matrix of pair experimental data that his been processed by the \code{pair} function.}

\item{n_assigned}{A matrix of pair experimental data cluster sizes that has been processed by the \code{pair} function.}
}
\description{
This function is used to impute potential outcomes for the \code{p_loop} function. 
Leaves out each pair and imputes its potential outcomes using random forests
on the remaining pairs. Individuals are treated as units when making predictions.
}
