% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loop.est.obj.R
\name{print.loopEst}
\alias{print.loopEst}
\title{Print loopEst object}
\usage{
\method{print}{loopEst}(
  x,
  digits = max(3L, getOption("digits")),
  signif.stars = getOption("show.signif.stars"),
  ...
)
}
\arguments{
\item{x}{loopEst object}
}
\value{
x
}
\description{
Print loopEst object
}
