% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loop.est.obj.R
\name{test.loopEst}
\alias{test.loopEst}
\title{Two-sided hypotheis test for LOOP estimate}
\usage{
test.loopEst(x, mu0 = 0, ...)
}
\arguments{
\item{x}{LOOP estimate}

\item{mu0}{null value for test (default=0)}
}
\description{
Two-sided hypotheis test for LOOP estimate
}
